refactor(queryHistory): filter when a query is provided
