+++ src/query/controllers/index.ts
@@ -1,13 +1,12 @@
 import { Controller, Get, HttpCode, Query } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { QueryService } from '../services';
-import { QueryResultDto } from '../dtos/queryResult.dto';
+import { QueryResultDto, GetQueriesDto } from '../dtos/';
 import { ApiInternalServerErrorMessageResponse } from 'src/common/decorators';
 import { JwtAccessToken, JwtRequestContext } from 'src/auth/decorators';
 import { JwtContext } from 'src/auth/interfaces';
 import { PaginationResponse } from 'src/common/dtos';
 import { ApiPaginationResponse } from 'src/common/decorators';
-import { GetPaginationsParamsDto } from '../../common/dtos';
 
 @ApiTags('Query')
 @Controller('queries')
@@ -21,7 +20,7 @@ export class QueyController {
   @Get('/')
   getQueries(
     @JwtRequestContext() context: JwtContext,
-    @Query() queryParams: GetPaginationsParamsDto,
+    @Query() queryParams: GetQueriesDto,
   ): Promise<PaginationResponse<QueryResultDto>> {
     return this.queryService.getQueries(context, queryParams);
   }
