+++ src/query/dtos/index.ts
@@ -1,0 +1,2 @@
+export { GetQueriesDto } from './getQueries.dto';
+export { QueryResultDto } from './queryResult.dto';
